refactor commands into Command classes
