+++ commands/__init__.py
@@ -1,0 +1,13 @@
+__all__ = [
+    "bio_comm",
+    "bloodline_comm",
+    "bloodline_members_comm",
+    "command",
+    "culture_comm",
+    "dynasty_comm",
+    "load_comm",
+    "person_comm",
+    "religion_comm",
+    "title_comm",
+    "tree_comm"
+]
